feat(assembler): add RG register aliases
The bit-twiddling opcodes have an interesting behavior where you can
only specify registers R0-R3, but R3 is special and not really R3.

It is referred to as RS in the manual, I guess as a generic placeholder
for IN/OUT, with knowledge of IOPos, so this isn't quite IN/OUT
registers in page 0.

Allow the assembler to recognize the RS alias directly and convert to
"R3", so that it's clearer what is meant